+++ zerver/lib/home.py
@@ -1,17 +1,14 @@
 import calendar
-import datetime
-import os
 import time
 from dataclasses import dataclass
 from typing import Any, Dict, List, Optional, Tuple
 
-import pytz
 from django.conf import settings
 from django.http import HttpRequest
 from django.utils import translation
-from django.utils.timezone import now as timezone_now
 from two_factor.utils import default_device
 
+from zerver.lib.compatibility import is_outdated_server
 from zerver.lib.events import do_events_register
 from zerver.lib.i18n import (
     get_and_set_request_language,
@@ -37,41 +34,6 @@ class UserPermissionInfo:
     is_realm_admin: bool
     is_realm_owner: bool
     show_webathena: bool
-
-
-# LAST_SERVER_UPGRADE_TIME is the last time the server had a version deployed.
-if settings.PRODUCTION:  # nocoverage
-    timestamp = os.path.basename(os.path.abspath(settings.DEPLOY_ROOT))
-    LAST_SERVER_UPGRADE_TIME = datetime.datetime.strptime(timestamp, "%Y-%m-%d-%H-%M-%S").replace(
-        tzinfo=pytz.utc
-    )
-else:
-    LAST_SERVER_UPGRADE_TIME = timezone_now()
-
-
-def is_outdated_server(user_profile: Optional[UserProfile]) -> bool:
-    # Release tarballs are unpacked via `tar -xf`, which means the
-    # `mtime` on files in them is preserved from when the release
-    # tarball was built.  Checking this allows us to catch cases where
-    # someone has upgraded in the last year but to a release more than
-    # a year old.
-    git_version_path = os.path.join(settings.DEPLOY_ROOT, "version.py")
-    release_build_time = datetime.datetime.utcfromtimestamp(
-        os.path.getmtime(git_version_path)
-    ).replace(tzinfo=pytz.utc)
-
-    version_no_newer_than = min(LAST_SERVER_UPGRADE_TIME, release_build_time)
-    deadline = version_no_newer_than + datetime.timedelta(
-        days=settings.SERVER_UPGRADE_NAG_DEADLINE_DAYS
-    )
-
-    if user_profile is None or not user_profile.is_realm_admin:
-        # Administrators get warned at the deadline; all users 30 days later.
-        deadline = deadline + datetime.timedelta(days=30)
-
-    if timezone_now() > deadline:
-        return True
-    return False
 
 
 def get_furthest_read_time(user_profile: Optional[UserProfile]) -> Optional[float]:
